Rendergraph hacky mvp
Fixed the bug with the compute shader. The issue was that the images weren't created by the rendergraph using the correct layout (they all used VK_LAYOUT_UNDEFINED), i'm still not sure why this error seemed to cause the visible light buffer to be zerod out after the compute shader invocations when trying to debug in renderdoc.